diary entries added to DB
