+++ amplify/backend/api/flow/parameters.json
@@ -1,0 +1,5 @@
+{
+    "AppSyncApiName": "flow",
+    "DynamoDBBillingMode": "PAY_PER_REQUEST",
+    "DynamoDBEnableServerSideEncryption": "false"
+}
