+++ amplify/backend/api/flow/schema.graphql
@@ -1,0 +1,5 @@
+type DiaryEntry @model {
+  id: ID!
+  title: String!
+  body: String
+}
