+++ amplify/backend/api/flow/transform.conf.json
@@ -1,0 +1,4 @@
+{
+    "Version": 5,
+    "ElasticsearchWarning": true
+}
